first review of setup.py. add back kalliope.py. THIS DOES NOT WORK
