+++ kalliope/_version.py
@@ -1,0 +1,3 @@
+# https://www.python.org/dev/peps/pep-0440/
+# version based on epoch
+version_str = "2016.11.20"
